==> Tutorials/00_Introduction.Rmd <==
---
title: "Ecological niche modeling based on the Mahalanobis distance and occurrence data"
author: "Laura Jiménez and Carola Franzen"
date: "July 2021"
output: html_document
fontsize: 12pt
linkcolor: dodgerblue4
bibliography: ENM-references.bib
csl: ecology-letters.csl
---

## Introduction

Our package contains all the functions needed to estimate the climatic fundamental niche of a species in a multivariate space defined by two or three environmental variables. We included three different statistical models to estimate the fundamental niche:

1) Mahalanobis model, which assumes that occurrences are observed according to a normal distribution and that ellipses are biologically realistic model to represent the border of the fundamental niche. Under this approach, it is implicitly assumed that every environmental combination in niche space is equally likely to be observed in a sample of occurrences.

2) Weigthed-normal model *sensu* @Jimenez2021, this model has the same assumptions as the Mahalanobis model regarding the shape of the fundamental niche, however, it takes into account that not all the environmental combinations in niche space have the same probability of being observed in a sample of occurrences. 

3) Bayesian normal model *sensu* @Jimenez2019, which takes the Mahalanobis as the underlying models for the fundamental niche of a species and assumes that there is relevant *a priori* information that can be incorporated into de model, usually from physiological experiments.

Additionally, we included the functions used to apply the evaluation method based on a hypergeometric test presented by @Jimenez2020.

The main goal of this tutorial is to provide worked examples that teach you how to use all the functions included in the package. Although we chose two arbitrary species as our focus species (a hummingbird, *Threnetes ruckeri*, and a butterfly, *Catasticta nimbice*), you should be able to apply all the functions to any species and any geographic region. In the following section, we provide a detailed description of the datasets that are included in the package and that can be used to reproduce all the examples, as well as to prepare other datasets for future analysis with other species and study regions.

### Organization within the tutorial

In each section of the tutorial, you will find a section called **Input data** which will indicate which are the files that you need beforehand to be able to reproduce the examples, and, a section called **Output data** which indicates what are the files that you would generate after reproducing the examples. In some cases, we only show how to make the analysis for one of the species and ask you to repeat the analysis for the second species as an exercise. It is important that you generate all the files listed in the **Output data** section since they will be used in subsequent sections of the tutorial.

Here is a list of the different sections within the tutorial and the corresponding functions of the package that are used in that specific section; the input and output files are also listed.

```{r organization, echo=FALSE}
nicetable <- read.csv("./Images/table_tutorials.csv",header=T)
knitr::kable(nicetable)
```

### References
