==> Tutorials/06_rsampleM_inE.Rmd <==
---
title: "Random sampling in environmental data"
author: "Laura Jiménez and Carola Franzen"
date: "June 2021"
output: html_document
fontsize: 12pt
linkcolor: dodgerblue4
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      warning = FALSE)
```
\

## Sampling random points from an area of interest

The function `rs.inE` takes a random sample of environmental combinations inside the study area. For this, the environmental data from a rasterstack is clipped by a shapefile that delimits the area before the sample is taken.

```{r, eval=FALSE}
rs.inE(region, N, Estck)
```
\

### Parameters


- `region` = a shapefile of the study area (polygon)

- `N` = the sample size

- `Estck` = a rasterstack that contains at least two layers with environmental data
\
\

### Input and Output

As input, rasters with environmental information and a shapefile that defines the region of interest are needed. 
The output of the function is a matrix of random points with as many columns as layers the rasterstack has.
\
\

### Dependencies


**Parameters:** `raster`, `rgdal`

**Example plot:** `scales`
\

\

## Worked Examples

### Read source code and libraries

```{r}
source(".\\Functions\\Trandom_sampling_inE.R")
```


```{r}
library(raster)
library(rgdal)
library(scales)
```
\


### *Catasticta nimbice*
\

**Input Data**
\

- bio1.asc

- bio12.asc

- Catasticta_nimbice.shp (and associated files)
\
\

First, the rasterstack for the parameter `Estck` is prepared. For this example two rasters with the environmental data "annual mean temperature" and "total annual precipitation" are used. The rasterstack will therefore have two layers.
```{r}
# Read environmental layers
bio1 <- raster("./Initial_Data/bio1.asc")
bio12 <- raster("./Initial_Data/bio12.asc")

# Create a single raster with as many layers as environmental variables
stck_bios <- stack(bio1, bio12)
```
\

Read the parameter for `region`.
```{r, results='hide', message=FALSE}
cn.shp <- readOGR("./Initial_Data/shapefiles","Catasticta_nimbice")
```
\

Apply the function with `Estck`, `region`, and `N`. Here, `N` will be 10,000. The generated matrix will therefore have two columns with environmental data from the study area and 10,000 rows.
```{r}
sam.pnts <- rs.inE(region = cn.shp, N = 10000, Estck = stck_bios)
```
\

If needed, save the output as a csv-file.
```{r, eval=FALSE}
write.csv(sam.pnts,file=paste0("./Generated_Data/Catasticta_nimbice","_M_E.csv"),
          row.names = F)
```
\

Create a simple plot of the environmental space (E-Space) with a species' occurrence.
```{r, fig.show='hide'}
# read occurrence data
occ <- read.csv("./Generated_Data/Catasticta_nimbice_occ_GE.csv",header=T) 
# set colors
pal <- c("grey50", "turquoise")

# plot
# x11()
#   use random points with environmental data for the plot
plot(sam.pnts[,1], sam.pnts[,2], pch=".", col=pal[1], 
     xlab="Mean Annual Temperature (°C*10)", ylab="Annual precipitation (mm)", 
     main="Environmental Space") 
# add points of species environmental data
points(occ$bio1, occ$bio12, pch=19, col=pal[2])
# add legend
legend(x= "topleft",
       legend = c("Study Area", "Occurences"),
       pch = c(20, 19),
       col = c(pal[1], pal[2]),
       bty = "n")
```
\
```{r, echo=FALSE, fig.cap=" Figure of the environmental space, where the occurrence points of *Catasticta nimbice* are plotted over a bunch of background points that were sampled from the rasterstack and polygon.", out.width = '70%',fig.align="center"}
options(knitr.duplicate.label = "allow")
knitr::include_graphics("Images/SampleM_Espace.png")
```
\


### *Threnetes ruckeri*
\

**Input Data**
\

- bio1.asc

- bio12.asc

- Threnetes_ruckeri.shp (and associated files)
\
\

In this example, a different graph will be plotted using the sampled points.

Read the shapefile. The rasterstack was already prepared in the previous example and it is not necessary to redo the process.
```{r, results='hide', message=FALSE}
# Read M polygon
thr.shp <- readOGR("./Initial_Data/shapefiles","Threnetes_ruckeri")

```
\

Apply the function with `region`, `Estck`, and a sample size with `N`.
```{r}
## Get a random sample of points in M and extract its corresponding environmental values
sam.pnts2 <- rs.inE(region = thr.shp,N = 10000,Estck = stck_bios)

```
\

The output is used to plot a graph showing concentrations of sample points marked by a shade. The layers of shades are created by kernel smoothing. The kernel is drawn using the points inside M. There are currently five levels of shades selected (5, 25, 50, 75, and 95). 
<!-- what does "probability of selecting a point" here mean? --> 
```{r, fig.show='hide'}
# Calculate and draw the kernel using the points inside M (define M carefully)
# kernel for contour plot
fhat.M <- ks::kde(x=cbind(sam.pnts2[,1], sam.pnts2[,2]))

## Probability of selecting a point, given the fundamental niche
# plot estimated kernel of M with points inside M
lvls1 <- c(5,25,50,75,95)
# library(scales)
col.M2 <- alpha("orangered2",0.1)
M.cols <- colorRampPalette(c(alpha("white",0.1),alpha("cadetblue4",0.9)))

# plot
# x11()
plot(fhat.M,display="filled.contour",cont=lvls1,main="",xlab="Mean annual temperature (°C*10)",
     ylab="Annual precipitation (mm)",col=M.cols(length(lvls1)+1))
# add points used for kernel estimation
points(sam.pnts2[,1], sam.pnts2[,2],col=col.M2,pch=20,cex=0.6)

```

```{r, echo=FALSE, fig.cap=" Figure of the environmental space showing layers with various density of the sampled points from the rasterstack and polygon.", out.width = '70%',fig.align="center"}
options(knitr.duplicate.label = "allow")
knitr::include_graphics("Images/ThR_Kernel_shade.png")
```
\
\

*End of tutorial*
